[RELEASE]: miniapp 0512

+++ packages/build-plugin-rax-app/src/plugins/miniapp/CopyJsx2mpRuntime.js
@@ -1,5 +1,5 @@
 const { join } = require('path');
-const { copySync, writeFileSync, ensureFileSync, readJSONSync, readFileSync } = require('fs-extra');
+const { copySync, existsSync, writeFileSync, readJSONSync, readFileSync } = require('fs-extra');
 const { minify } = require('terser');
 
 /**
@@ -48,12 +48,14 @@ module.exports = class JSX2MPRuntimePlugin {
           : runtimePackageJSON.main || 'index.js';
         const sourceFile = require.resolve(join(runtimePackagePath, runtimeTargetPath));
         const targetFile = join(this.outputPath, 'npm', runtime + '.js');
-        ensureFileSync(targetFile);
+
         if (this.mode === 'build') {
           const sourceCode = minify(readFileSync(sourceFile, 'utf-8')).code;
           writeFileSync(targetFile, sourceCode);
         } else {
-          copySync(sourceFile, targetFile);
+          if (!existsSync(targetFile)) {
+            copySync(sourceFile, targetFile);
+          }
         }
 
         callback();
